post seed. do not delete img from img_seed folder - if delete data

+++ Forum/Forum.DataAccess/Repository/FileManager.cs
@@ -20,7 +20,9 @@ namespace Forum.DataAccess.Repository
         {
             string webRootPath = _hostEnvironment.WebRootPath;
             var imagePath = Path.Combine(webRootPath, filePath.TrimStart('\\'));
-            if (System.IO.File.Exists(imagePath))
+            bool asd = !imagePath.Contains("img_seed");
+            bool dsa = System.IO.File.Exists(imagePath) && !imagePath.Contains("img_seed");
+            if (System.IO.File.Exists(imagePath) && !imagePath.Contains("img_seed"))
             {
                 System.IO.File.Delete(imagePath);
             }
